Organized SectionSelect tests * also removed fetching from SectionSelect tests * added style assertions for excess enrollment
