+++ autoscheduler/frontend/src/tests/redux/Availability.test.ts
