+++ autoscheduler/frontend/src/tests/redux/Schedule.test.ts
