+++ autoscheduler/frontend/src/tests/ui/CourseSelectCard.test.tsx
@@ -17,11 +17,6 @@ import CourseSelectCard from '../../components/SchedulingPage/CourseSelectColumn
 import autoSchedulerReducer from '../../redux/reducer';
 import testFetch from '../testData';
 import setTerm from '../../redux/actions/term';
-import SectionSelect from '../../components/SchedulingPage/CourseSelectColumn/CourseSelectCard/ExpandedCourseCard/SectionSelect/SectionSelect';
-import Section from '../../types/Section';
-import Instructor from '../../types/Instructor';
-import Meeting, { MeetingType } from '../../types/Meeting';
-import { updateCourseCard } from '../../redux/actions/courseCards';
 
 const dummySectionArgs = {
   id: 123456,
@@ -440,110 +435,6 @@ describe('Course Select Card UI', () => {
     });
   });
 
-  describe('handles honors icon', () => {
-    test('by showing it for honors sections', async () => {
-      // arrange
-      fetchMock.mockResponseOnce(JSON.stringify({ // api/course/search
-        results: ['MATH 151'],
-      }));
-      fetchMock.mockImplementationOnce(testFetch); // api/sections
-
-      const store = createStore(autoSchedulerReducer, applyMiddleware(thunk));
-      store.dispatch(setTerm('201931'));
-      const {
-        getByText, getByLabelText, findByText, findByTestId,
-      } = render(
-        <Provider store={store}><CourseSelectCard id={0} /></Provider>,
-      );
-
-      // act
-      fireEvent.change(getByLabelText('Course'), { target: { value: 'M' } });
-      fireEvent.click(await findByText('MATH 151'));
-
-      // switch to sections view
-      fireEvent.click(getByText('Section'));
-      const honorsIcon = await findByTestId('honors');
-
-      // assert
-      expect(honorsIcon).toBeInTheDocument();
-    });
-
-    test('by hiding it for regular sections', async () => {
-      // arrange
-      fetchMock.mockResponseOnce(JSON.stringify({ // api/course/search
-        results: ['CSCE 121'],
-      }));
-      fetchMock.mockImplementationOnce(testFetch); // api/sections
-
-      const store = createStore(autoSchedulerReducer, applyMiddleware(thunk));
-      store.dispatch(setTerm('201931'));
-      const {
-        getByText, getByLabelText, findByText, queryByTitle,
-      } = render(
-        <Provider store={store}><CourseSelectCard id={0} /></Provider>,
-      );
-
-      // act
-      fireEvent.change(getByLabelText('Course'), { target: { value: 'C' } });
-      fireEvent.click(await findByText('CSCE 121'));
-
-      // switch to sections view
-      fireEvent.click(getByText('Section'));
-      await findByText((cont, el) => ignoreInvisible(cont, el, '501'));
-      const honorsIcon = queryByTitle('honors');
-
-      // assert
-      expect(honorsIcon).not.toBeInTheDocument();
-    });
-
-    test('if the same professor has both honors and regular sections', async () => {
-      // arrange
-      fetchMock.mockResponseOnce(JSON.stringify({ // api/course/search
-        results: ['MATH 151'],
-      }));
-      const math151Dummy = {
-        ...dummySectionArgs, subject: 'MATH', course_num: 151,
-      };
-      fetchMock.mockResponseOnce(JSON.stringify([ // api/sections
-        {
-          ...math151Dummy,
-          section_num: '201',
-          instructor_name: 'Aakash Tyagi',
-          honors: true,
-        },
-        {
-          ...math151Dummy,
-          section_num: '502',
-          instructor_name: 'Aakash Tyagi',
-          honors: false,
-        },
-      ]));
-
-      const store = createStore(autoSchedulerReducer, applyMiddleware(thunk));
-      store.dispatch(setTerm('201931'));
-      const {
-        getByText, getByLabelText, findByText, findAllByText,
-      } = render(
-        <Provider store={store}><CourseSelectCard id={0} /></Provider>,
-      );
-
-      // act
-      fireEvent.change(getByLabelText('Course'), { target: { value: 'M' } });
-      fireEvent.click(await findByText('MATH 151'));
-
-      // switch to sections view
-      fireEvent.click(getByText('Section'));
-      const profNames = await findAllByText('Aakash Tyagi');
-      expect(profNames).toHaveLength(2);
-      const honorsSection = profNames[0];
-      const regularSection = profNames[1];
-
-      // assert
-      expect(queryByTitleIn(honorsSection, 'Honors')).toBeInTheDocument();
-      expect(queryByTitleIn(regularSection, 'Honors')).not.toBeInTheDocument();
-    });
-  });
-
   describe('does not fetch inappropriately', () => {
     describe('when we search and go to the Sections tab', () => {
       test('and collapse then expand the card', async () => {
@@ -804,440 +695,6 @@ describe('Course Select Card UI', () => {
       // assert
       const placeholder = 'There are no available sections for this term';
       expect(queryByText(placeholder)).not.toBeInTheDocument();
-    });
-  });
-
-  describe('SectionSelect', () => {
-    describe('shows ONLINE', () => {
-      test('for 00:00 meeting times & online section', () => {
-        // arrange
-        const store = createStore(autoSchedulerReducer, applyMiddleware(thunk));
-
-        const {
-          getByText,
-        } = render(
-          <Provider store={store}><SectionSelect id={0} /></Provider>,
-        );
-
-        const testSection = new Section({
-          id: 0,
-          crn: 0,
-          subject: 'CSCE',
-          courseNum: '121',
-          sectionNum: '200',
-          minCredits: 3,
-          maxCredits: null,
-          currentEnrollment: 0,
-          maxEnrollment: 0,
-          instructor: new Instructor({
-            name: 'Test',
-          }),
-          honors: false,
-          web: true,
-          grades: null,
-        });
-        const testMeeting = new Meeting({
-          id: 1,
-          building: '',
-          meetingDays: new Array(7).fill(true),
-          startTimeHours: 0,
-          startTimeMinutes: 0,
-          endTimeHours: 0,
-          endTimeMinutes: 0,
-          meetingType: MeetingType.LEC,
-          section: testSection,
-        });
-
-        const sectionSelected = {
-          section: testSection,
-          meetings: [testMeeting],
-          selected: false,
-        };
-
-        // Add the SectionSelected type to the store so it shows up in the SectionSelect component
-        store.dispatch<any>(updateCourseCard(0, {
-          sections: [sectionSelected],
-        }, '201931'));
-
-        // assert
-        expect(getByText('ONLINE')).toBeTruthy();
-      });
-    });
-
-    describe('does not show ONLINE', () => {
-      test('for 00:00 meeting times that are not online sections', () => {
-        // arrange
-        const store = createStore(autoSchedulerReducer, applyMiddleware(thunk));
-
-        const {
-          queryByText,
-        } = render(
-          <Provider store={store}><SectionSelect id={0} /></Provider>,
-        );
-
-        const testSection = new Section({
-          id: 0,
-          crn: 0,
-          subject: 'CSCE',
-          courseNum: '121',
-          sectionNum: '200',
-          minCredits: 3,
-          maxCredits: null,
-          currentEnrollment: 0,
-          maxEnrollment: 0,
-          instructor: new Instructor({
-            name: 'Test',
-          }),
-          honors: false,
-          web: false,
-          grades: null,
-        });
-        const testMeeting = new Meeting({
-          id: 1,
-          building: '',
-          meetingDays: new Array(7).fill(true),
-          startTimeHours: 0,
-          startTimeMinutes: 0,
-          endTimeHours: 0,
-          endTimeMinutes: 0,
-          meetingType: MeetingType.LEC,
-          section: testSection,
-        });
-
-        const sectionSelected = {
-          section: testSection,
-          meetings: [testMeeting],
-          selected: false,
-        };
-
-        // Add the SectionSelected type to the store so it shows up in the SectionSelect component
-        store.dispatch<any>(updateCourseCard(0, {
-          sections: [sectionSelected],
-        }, '201931'));
-
-        // assert
-        expect(queryByText('ONLINE')).not.toBeInTheDocument();
-      });
-    });
-
-    describe('shows a single meeting time', () => {
-      test('for duplicate exam times', () => {
-      // arrange
-        const store = createStore(autoSchedulerReducer, applyMiddleware(thunk));
-
-        const { getAllByText } = render(
-          <Provider store={store}><SectionSelect id={0} /></Provider>,
-        );
-
-        const testSection = new Section({
-          id: 0,
-          crn: 0,
-          subject: 'CSCE',
-          courseNum: '121',
-          sectionNum: '200',
-          minCredits: 3,
-          maxCredits: null,
-          currentEnrollment: 0,
-          maxEnrollment: 0,
-          instructor: new Instructor({
-            name: 'Test',
-          }),
-          honors: false,
-          web: true,
-          grades: null,
-        });
-        const testMeeting = new Meeting({
-          id: 1,
-          building: '',
-          meetingDays: new Array(7).fill(true),
-          startTimeHours: 19,
-          startTimeMinutes: 0,
-          endTimeHours: 20,
-          endTimeMinutes: 0,
-          meetingType: MeetingType.EXAM,
-          section: testSection,
-        });
-
-        const sectionSelected = {
-          section: testSection,
-          meetings: [testMeeting, testMeeting, testMeeting],
-          selected: false,
-        };
-
-        // Add the SectionSelected type to the store so it shows up in the SectionSelect component
-        store.dispatch<any>(updateCourseCard(0, {
-          sections: [sectionSelected],
-        }, '201931'));
-
-        // assert
-        expect(getAllByText('EXAM')).toHaveLength(1);
-      });
-
-      test('for slightly different exam times, as long as they have the same start time', () => {
-        // arrange
-        const store = createStore(autoSchedulerReducer, applyMiddleware(thunk));
-
-        const { getAllByText } = render(
-          <Provider store={store}><SectionSelect id={0} /></Provider>,
-        );
-
-        const testSection = new Section({
-          id: 0,
-          crn: 0,
-          subject: 'CSCE',
-          courseNum: '121',
-          sectionNum: '200',
-          minCredits: 3,
-          maxCredits: null,
-          currentEnrollment: 0,
-          maxEnrollment: 0,
-          instructor: new Instructor({
-            name: 'Test',
-          }),
-          honors: false,
-          web: true,
-          grades: null,
-        });
-        const testMeeting = new Meeting({
-          id: 1,
-          building: '',
-          meetingDays: new Array(7).fill(true),
-          startTimeHours: 19,
-          startTimeMinutes: 0,
-          endTimeHours: 20,
-          endTimeMinutes: 30,
-          meetingType: MeetingType.EXAM,
-          section: testSection,
-        });
-        const testMeeting2 = {
-          ...testMeeting,
-          endTimeHours: 21,
-          endTimeMinutes: 0,
-        };
-
-        const sectionSelected = {
-          section: testSection,
-          meetings: [testMeeting, testMeeting, testMeeting, testMeeting2],
-          selected: false,
-        };
-
-        // Add the SectionSelected type to the store so it shows up in the SectionSelect component
-        store.dispatch<any>(updateCourseCard(0, {
-          sections: [sectionSelected],
-        }, '201931'));
-
-        // assert
-        expect(getAllByText('EXAM')).toHaveLength(1);
-      });
-
-      test('for labs on different days, as long as they have the same start time', () => {
-        // arrange
-        const store = createStore(autoSchedulerReducer, applyMiddleware(thunk));
-
-        const { getAllByText } = render(
-          <Provider store={store}><SectionSelect id={0} /></Provider>,
-        );
-
-        const testSection = new Section({
-          id: 0,
-          crn: 0,
-          subject: 'CSCE',
-          courseNum: '121',
-          sectionNum: '200',
-          minCredits: 3,
-          maxCredits: null,
-          currentEnrollment: 0,
-          maxEnrollment: 0,
-          instructor: new Instructor({
-            name: 'Test',
-          }),
-          honors: false,
-          web: true,
-          grades: null,
-        });
-        const testMeeting = new Meeting({
-          id: 1,
-          building: '',
-          meetingDays: [true, false, false, false, false, false, false],
-          startTimeHours: 19,
-          startTimeMinutes: 0,
-          endTimeHours: 20,
-          endTimeMinutes: 30,
-          meetingType: MeetingType.LAB,
-          section: testSection,
-        });
-        const testMeeting2 = {
-          ...testMeeting,
-          meetingDays: [false, true, false, false, false, false, false],
-          endTimeHours: 21,
-          endTimeMinutes: 0,
-        };
-
-        const sectionSelected = {
-          section: testSection,
-          meetings: [testMeeting, testMeeting, testMeeting, testMeeting2],
-          selected: false,
-        };
-
-        // Add the SectionSelected type to the store so it shows up in the SectionSelect component
-        store.dispatch<any>(updateCourseCard(0, {
-          sections: [sectionSelected],
-        }, '201931'));
-
-        // assert
-        expect(getAllByText('LAB')).toHaveLength(1);
-      });
-    });
-
-    describe('shows multiple meeting times', () => {
-      test('for different meeting types at different times', () => {
-        // arrange
-        const store = createStore(autoSchedulerReducer, applyMiddleware(thunk));
-
-        const { getAllByText } = render(
-          <Provider store={store}><SectionSelect id={0} /></Provider>,
-        );
-
-        const testSection = new Section({
-          id: 0,
-          crn: 0,
-          subject: 'CSCE',
-          courseNum: '121',
-          sectionNum: '200',
-          minCredits: 3,
-          maxCredits: null,
-          currentEnrollment: 0,
-          maxEnrollment: 0,
-          instructor: new Instructor({
-            name: 'Test',
-          }),
-          honors: false,
-          web: true,
-          grades: null,
-        });
-        const testMeeting1 = new Meeting({
-          id: 1,
-          building: '',
-          meetingDays: new Array(7).fill(true),
-          startTimeHours: 19,
-          startTimeMinutes: 0,
-          endTimeHours: 20,
-          endTimeMinutes: 0,
-          meetingType: MeetingType.EXAM,
-          section: testSection,
-        });
-        const testMeeting2 = {
-          ...testMeeting1,
-          startTimeHours: 9,
-          startTimeMinutes: 0,
-          endTimeHours: 10,
-          endTimeMinutes: 30,
-          meetingType: MeetingType.LEC,
-        };
-
-        const sectionSelected = {
-          section: testSection,
-          meetings: [testMeeting1, testMeeting2],
-          selected: false,
-        };
-
-        // Add the SectionSelected type to the store so it shows up in the SectionSelect component
-        store.dispatch<any>(updateCourseCard(0, {
-          sections: [sectionSelected],
-        }, '201931'));
-
-        // assert
-        expect(getAllByText(/(EXAM)|(LEC)/)).toHaveLength(2);
-      });
-    });
-  });
-
-  describe('displays correct number of seats', () => {
-    test('when there are no students enrolled', async () => {
-      // arrange
-      fetchMock.mockResponseOnce(JSON.stringify({ // api/course/search
-        results: ['CSCE 121', 'CSCE 221', 'CSCE 312'],
-      }));
-      fetchMock.mockImplementationOnce(testFetch); // api/sections
-
-      const store = createStore(autoSchedulerReducer, applyMiddleware(thunk));
-      store.dispatch(setTerm('201931'));
-      const {
-        getByLabelText, getByText, findByText, findAllByText,
-      } = render(
-        <Provider store={store}><CourseSelectCard id={0} /></Provider>,
-      );
-
-      // act
-      // fill in course and go to section view
-      const courseEntry = getByLabelText('Course') as HTMLInputElement;
-      fireEvent.change(courseEntry, { target: { value: 'CSCE ' } });
-      fireEvent.click(await findByText('CSCE 121'));
-
-      fireEvent.click(getByText('Section'));
-
-      // assert
-      // verify text for seats is correct
-      const sectionText = (await findAllByText(/ seats left/))[0];
-      expect(sectionText).toHaveTextContent('25/25 seats left');
-    });
-
-    test('when a section has max enrollment', async () => {
-      // arrange
-      fetchMock.mockResponseOnce(JSON.stringify({ // api/course/search
-        results: ['CSCE 121', 'CSCE 221', 'CSCE 312'],
-      }));
-      fetchMock.mockImplementationOnce(testFetch); // api/sections
-
-      const store = createStore(autoSchedulerReducer, applyMiddleware(thunk));
-      store.dispatch(setTerm('201931'));
-      const {
-        getByLabelText, getByText, findByText, findAllByText,
-      } = render(
-        <Provider store={store}><CourseSelectCard id={0} /></Provider>,
-      );
-
-      // act
-      // fill in course and go to section view
-      const courseEntry = getByLabelText('Course') as HTMLInputElement;
-      fireEvent.change(courseEntry, { target: { value: 'CSCE ' } });
-      fireEvent.click(await findByText('CSCE 121'));
-
-      fireEvent.click(getByText('Section'));
-
-      // assert
-      // verify text for seats is correct
-      const sectionText = (await findAllByText(/ seats left/))[1];
-      expect(sectionText).toHaveTextContent('0/25 seats left');
-    });
-
-    test('when a section has over max enrollement', async () => {
-      // arrange
-      fetchMock.mockResponseOnce(JSON.stringify({ // api/course/search
-        results: ['CSCE 121', 'CSCE 221', 'CSCE 312'],
-      }));
-      fetchMock.mockImplementationOnce(testFetch); // api/sections
-
-      const store = createStore(autoSchedulerReducer, applyMiddleware(thunk));
-      store.dispatch(setTerm('201931'));
-      const {
-        getByLabelText, getByText, findByText, findAllByText,
-      } = render(
-        <Provider store={store}><CourseSelectCard id={0} /></Provider>,
-      );
-
-      // act
-      // fill in course and go to section view
-      const courseEntry = getByLabelText('Course') as HTMLInputElement;
-      fireEvent.change(courseEntry, { target: { value: 'CSCE ' } });
-      fireEvent.click(await findByText('CSCE 121'));
-
-      fireEvent.click(getByText('Section'));
-
-      // assert
-      // verify text for seats is correct
-      const sectionText = (await findAllByText(/ seats left/))[2];
-      expect(sectionText).toHaveTextContent('-1/25 seats left');
     });
   });
 });
